refactor: changed directory - moved config, validate to utils
